# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Grainily API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/branding:
    get:
      operationId: getBranding
      tags: [admin]
      summary: Get current branding settings
      responses:
        "200":
          description: Branding settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Branding"
    put:
      operationId: updateBranding
      tags: [admin]
      summary: Update branding settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Branding"
      responses:
        "200":
          description: Updated branding
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Branding"
  /admin/categories:
    get:
      operationId: adminListCategories
      tags: [admin]
      summary: Admin - list all categories
      responses:
        "200":
          description: All categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
    post:
      operationId: adminCreateCategory
      tags: [admin]
      summary: Admin - create category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryInput"
      responses:
        "201":
          description: Created category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
  /admin/categories/{id}:
    put:
      operationId: adminUpdateCategory
      tags: [admin]
      summary: Admin - update category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryInput"
      responses:
        "200":
          description: Updated category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
    delete:
      operationId: adminDeleteCategory
      tags: [admin]
      summary: Admin - delete category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /admin/plans:
    get:
      operationId: adminListPlans
      tags: [admin]
      summary: Admin - list all plans
      responses:
        "200":
          description: All plans
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Plan"
    post:
      operationId: adminCreatePlan
      tags: [admin]
      summary: Admin - create plan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlanInput"
      responses:
        "201":
          description: Created plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Plan"
  /admin/plans/{id}:
    put:
      operationId: adminUpdatePlan
      tags: [admin]
      summary: Admin - update plan
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlanInput"
      responses:
        "200":
          description: Updated plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Plan"
    delete:
      operationId: adminDeletePlan
      tags: [admin]
      summary: Admin - delete plan
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /admin/config/{key}:
    put:
      operationId: adminUpdateConfig
      tags: [admin]
      summary: Admin - set a platform config value
      parameters:
        - name: key
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlatformConfigInput"
      responses:
        "200":
          description: Updated config
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlatformConfig"
...
components:
  schemas:
...
    Branding:
      type: object
      properties:
        logoUrl:
          type: ["string", "null"]
        primaryColor:
          type: string
        accentColor:
          type: string
        backgroundColor:
          type: string
        fontFamily:
          type: string
        siteName:
          type: string
    Category:
      type: object
      required: [id, name, slug, active]
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        active:
          type: boolean
    CategoryInput:
      type: object
      required: [name, slug]
      properties:
        name:
          type: string
        slug:
          type: string
        active:
          type: boolean
...
    Plan:
      type: object
      required: [id, name, slug, priceMonthly, features, isActive, sortOrder]
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        description:
          type: ["string", "null"]
        priceMonthly:
          type: string
        priceAnnual:
          type: ["string", "null"]
        maxAds:
          type: ["integer", "null"]
        features:
          type: array
          items:
            type: string
        isActive:
          type: boolean
        sortOrder:
          type: integer
    PlanInput:
      type: object
      required: [name, slug, priceMonthly, features]
      properties:
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        priceMonthly:
          type: string
        priceAnnual:
          type: string
        maxAds:
          type: integer
        features:
          type: array
          items:
            type: string
        isActive:
          type: boolean
        sortOrder:
          type: integer
    PlatformConfig:
      type: object
      required: [key, isSecret, label]
      properties:
        key:
          type: string
        value:
          type: ["string", "null"]
        isSecret:
          type: boolean
        label:
          type: string
        description:
          type: ["string", "null"]
    PlatformConfigInput:
      type: object
      required: [value]
      properties:
        value:
          type: string
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT